requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
bcrypt==4.0.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
deprecation==2.1.0
ecdsa==0.19.0
fastapi==0.115.0
frozenlist==1.4.1
gotrue==2.9.2
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
packaging==24.1
passlib==1.7.4
pathlib==1.0.1
pluggy==1.5.0
postgrest==0.17.1
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
PyYAML==6.0.2
realtime==2.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.1
yarl==1.15.2
